+++ requirements.txt
@@ -1,3 +1,5 @@
 openai
 exifread
 requests
+neo4j
+python-dotenv
